Reverse openid flow
Let users first log in on an openid provider, and then present a list of
zones to which they can log in. The login provider chooser can
eventually be skipped based on the host header if multiple dns names are
set up. The zone chooser should eventually be limited to the zones the
user has access to and could be skipped if there is only a single zone.

Signed-off-by: Peter Verraedt <peter@verraedt.be>

Skip init_and_secure_views for zone chooser

Signed-off-by: Peter Verraedt <peter@verraedt.be>

Fix irods_zones_config

Signed-off-by: Peter Verraedt <peter@verraedt.be>

Fix redirect url

Signed-off-by: Peter Verraedt <peter@verraedt.be>

Fix form action url

Signed-off-by: Peter Verraedt <peter@verraedt.be>

Fix order of assignments

Signed-off-by: Peter Verraedt <peter@verraedt.be>