separate bulk operations as macro
This commit separates the bulk operations in a macro, so they can
easily imported in different templates which represent a collection's
contents (collection view, trash, upcoming gallery view).

This commit also changes the bulk operations so the checkboxes
are identified by a class or ID, rather than their position in the
template. This will hopefully facilitate new ways of representing
collections, with slightly different structures.